requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
bottle==0.13.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
distro==1.9.0
editor==1.6.6
exceptiongroup==1.2.2
ffmpeg==1.4
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama_stack==0.0.1a5
llvmlite==0.43.0
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
narwhals==1.14.3
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
ollama==0.4.2
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.0
proxy-tools==0.1.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-Security==10.3.2
pyobjc-framework-WebKit==10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywebview==5.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
runs==1.2.2
semantic-version==2.10.0
setuptools-rust==1.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xmod==1.8.1
zipp==3.21.0
